server/requirements.txt
fastapi (>=0.135.1,<0.136.0)
uvicorn
starlette (>=0.52.1,<0.53.0)
sqlalchemy (>=2.0.48,<3.0.0)
python-jose (>=3.5.0,<4.0.0)
pydantic-settings (>=2.13.1,<3.0.0)
asyncpg
alembic (>=1.18.4,<2.0.0)
passlib==1.7.4
bcrypt==4.1.2
pydantic[email] (>=2.12.5,<3.0.0)
python-multipart (>=0.0.22,<0.0.23)
python-dateutil (>=2.9.0.post0,<3.0.0)
python-docx (>=1.2.0,<2.0.0)
